+++ tests/unit/utils/bytecode_tools/bytecodeDisassembler.spec.ts
@@ -20,10 +20,10 @@
 
 
 import { describe, test, expect } from 'vitest'
-import {Utils} from '@/utils/bytecode_tools/disassembler/Utils/index.ts'
+import { Helpers } from '@/utils/bytecode_tools/disassembler/utils/helpers.ts'
 import { DISASSEMBLER_TESTING_ASSETS } from './bytecode_assets';
-import {EVM_OPCODES} from '@/utils/bytecode_tools/disassembler/evm_opcodes/index.ts'
-import {Disassembler} from '@/utils/bytecode_tools/disassembler/BytecodeDisassembler.ts'
+import { EVM_OPCODES } from '@/utils/bytecode_tools/disassembler/utils/evm_opcodes.ts'
+import { Disassembler } from '@/utils/bytecode_tools/disassembler/BytecodeDisassembler.ts'
 
 describe('EVM Bytecode Disassembler Tests', () => {
   describe('Utility functions tests', () => {
@@ -31,7 +31,7 @@ describe('EVM Bytecode Disassembler Tests', () => {
       const BYTECODE = '  0x6080604052603e80600f5f395ff3fe ';
       const EXPECTED_SANITIZED_BYTECODE = '6080604052603e80600f5f395ff3fe';
 
-      const sanitizedBytecode = Utils.prepBytecode(BYTECODE);
+      const sanitizedBytecode = Helpers.prepBytecode(BYTECODE);
 
       expect(sanitizedBytecode).to.eq(EXPECTED_SANITIZED_BYTECODE);
     });
@@ -39,7 +39,7 @@ describe('EVM Bytecode Disassembler Tests', () => {
     test('Should return null if bytecode has odd length', () => {
       const BYTECODE = '0x40F';
 
-      expect(Utils.prepBytecode(BYTECODE)).to.be.null;
+      expect(Helpers.prepBytecode(BYTECODE)).to.be.null;
     });
 
     test('Should get operands for opcodes in PUSH family', () => {
@@ -51,7 +51,7 @@ describe('EVM Bytecode Disassembler Tests', () => {
       //         Therefore, offset will skip the first byte of the opcodes and 4 bytes of the operands
       const expectedOffset = index + 4 * 2;
 
-      const operandObject = Utils.getPushOperands(opcode!, bytecode, index);
+      const operandObject = Helpers.getPushOperands(opcode!, bytecode, index);
 
       const expectedOperand = ['80', '40', '1a', 'bf'];
 
@@ -70,7 +70,7 @@ describe('EVM Bytecode Disassembler Tests', () => {
 
       const expectedOperand = ['80', '40', '1a', 'bf'];
 
-      const result = Utils.parseBytecode(bytecode, byte, index);
+      const result = Helpers.parseBytecode(bytecode, byte, index);
 
       expect(result.index).to.eq(expectedOffset);
       expect(result.opcodeRepresentation.hex).to.eq(byte);
@@ -86,7 +86,7 @@ describe('EVM Bytecode Disassembler Tests', () => {
       const index = 8;
       const byte = bytecode.substring(index, index + 2); // MSTORE
 
-      const result = Utils.parseBytecode(bytecode, byte, index);
+      const result = Helpers.parseBytecode(bytecode, byte, index);
 
       expect(result.index).to.eq(index);
       expect(result.opcodeRepresentation.hex).to.eq(byte);
@@ -100,7 +100,7 @@ describe('EVM Bytecode Disassembler Tests', () => {
       const index = 8;
       const byte = bytecode.substring(index, index + 2); // Invalid opcode
 
-      const result = Utils.parseBytecode(bytecode, byte, index);
+      const result = Helpers.parseBytecode(bytecode, byte, index);
 
       expect(result.index).to.eq(index);
       expect(result.opcodeRepresentation.hex).to.eq(byte);
@@ -116,7 +116,7 @@ describe('EVM Bytecode Disassembler Tests', () => {
       const EXPECTED_BYCODE =
         '0x608060405234801561000f575f80fd5b506101438061001d5f395ff3fe608060405234801561000f575f80fd5b5060043610610034575f3560e01c80636b06623714610038578063e69f19cd14610054575b5f80fd5b610052600480360381019061004d91906100ba565b610072565b005b61005c61007b565b60405161006991906100f4565b60405180910390f35b805f8190555050565b5f8054905090565b5f80fd5b5f819050919050565b61009981610087565b81146100a3575f80fd5b50565b5f813590506100b481610090565b92915050565b5f602082840312156100cf576100ce610083565b5b5f6100dc848285016100a6565b91505092915050565b6100ee81610087565b82525050565b5f6020820190506101075f8301846100e5565b9291505056fe';
 
-      const noMetadataBycode = Utils.removeCBORMetadata(BYTECODE);
+      const noMetadataBycode = Helpers.removeCBORMetadata(BYTECODE);
 
       expect(noMetadataBycode).to.eq(EXPECTED_BYCODE);
     });
@@ -124,7 +124,7 @@ describe('EVM Bytecode Disassembler Tests', () => {
     test('Should leave the bytecode as-is if there is no CBOR-encoded metadata', () => {
       const BYTECODE = '0x6080604052';
 
-      const noMetadataBycode = Utils.removeCBORMetadata(BYTECODE);
+      const noMetadataBycode = Helpers.removeCBORMetadata(BYTECODE);
 
       expect(noMetadataBycode).to.eq(BYTECODE);
     });
